general refactors around timers and settings
