+++ src/components/TimerFace.vue
@@ -1,10 +1,16 @@
 <template>
-  <div>
+  <v-card>
+    <v-card-title class="headline">
+      {{ duration.title }}
+    </v-card-title>
+    <v-divider />
     <div class="timer-face my-4 display-4">
-      {{ formattedDuration(countdown) }}
+      {{ formattedDuration }}
     </div>
     <v-progress-linear v-model="progress" height="16" />
-  </div>
+    <v-divider />
+    <slot></slot>
+  </v-card>
 </template>
 
 <script>
@@ -14,35 +20,42 @@ import { timeFormat } from '../utils/helpers';
 
 export default {
   name: 'TimerFace',
-  props: {
-    duration: {
-      type: Number,
-      required: true,
-    }
-  },
   data() {
     return {
+      duration: {
+        title: '',
+        duration: 0
+      },
       progress: 0,
       countdown: 0
     };
   },
   computed: {
     ...mapGetters([
-      'getTimerValue'
+      'getActiveTimer', 'getTimerById', 'getTimerValue'
     ]),
     formattedDuration() {
-      return countdown => timeFormat(countdown);
+      return timeFormat(this.countdown);
     },
   },
   watch: {
+    getActiveTimer(newTimerId) {
+      const duration = this.getTimerById(newTimerId);
+      this.duration = duration;
+      this.progress = 0;
+    },
     getTimerValue(newTime) {
       this.countdown = newTime;
-      const progress = (this.duration - newTime) / this.duration;
+      const progress = (this.duration.duration - newTime) / this.duration.duration;
       this.progress = progress * 100;
     }
   },
   created() {
-    this.countdown = this.duration;
+    const duration = this.getTimerById(this.getActiveTimer);
+    const progress = (duration.duration - this.getTimerValue) / duration.duration;
+    this.duration = duration;
+    this.progress = progress * 100;
+    this.countdown = duration.duration;
   }
 };
 </script>
