+++ src/store/modules/settings/actions.js
@@ -1,6 +1,4 @@
 import {
-  SET_AUTOPLAY,
-  SET_CYCLE,
   SET_PLAY_CHIME,
   SET_CHIME,
   SET_STATS_WINDOW
@@ -11,16 +9,6 @@ import { Storage } from '../../../utils/storage';
 const settingsStorage = new Storage('settings');
 
 export const actions = {
-  setAutoplay(store, payload) {
-    store.commit(SET_AUTOPLAY, payload);
-
-    settingsStorage.add({ autoplay: payload });
-  },
-  setCycle(store, payload) {
-    store.commit(SET_CYCLE, payload);
-
-    settingsStorage.add({ cycle: payload });
-  },
   setPlayChime(store, payload) {
     store.commit(SET_PLAY_CHIME, payload);
 
@@ -40,14 +28,6 @@ export const actions = {
     const settings = settingsStorage.load();
 
     if (settings) {
-      if ('autoplay' in settings) {
-        store.commit(SET_AUTOPLAY, settings.autoplay);
-      }
-
-      if ('cycle' in settings) {
-        store.commit(SET_CYCLE, settings.cycle);
-      }
-
       if ('play_chime' in settings) {
         store.commit(SET_PLAY_CHIME, settings.play_chime);
       }
