+++ src/store/modules/settings/getters.js
@@ -1,6 +1,4 @@
 export const getters = {
-  getAutoplay: state => state.autoplay,
-  getCycle: state => state.cycle,
   getChime: state => state.chime,
   getPlayChime: state => state.play_chime,
   getStatsWindow: state => state.stats_window
