+++ src/store/modules/settings/mutation_types.js
@@ -1,5 +1,3 @@
-export const SET_AUTOPLAY = 'SET_AUTOPLAY';
-export const SET_CYCLE = 'SET_CYCLE';
 export const SET_PLAY_CHIME = 'SET_PLAY_CHIME';
 export const SET_CHIME = 'SET_CHIME';
 export const SET_STATS_WINDOW = 'SET_STATS_WINDOW';
