+++ src/store/modules/settings/mutations.js
@@ -1,20 +1,12 @@
 import Vue from 'vue';
 
 import {
-  SET_AUTOPLAY,
-  SET_CYCLE,
   SET_PLAY_CHIME,
   SET_CHIME,
   SET_STATS_WINDOW
 } from './mutation_types';
 
 export const mutations = {
-  [SET_AUTOPLAY](state, payload) {
-    Vue.set(state, 'autoplay', payload);
-  },
-  [SET_CYCLE](state, payload) {
-    Vue.set(state, 'cycle', payload);
-  },
   [SET_PLAY_CHIME](state, payload) {
     Vue.set(state, 'play_chime', payload);
   },
